refactor(common)!: dissolve org.apache.hudi.util grab-bag in hudi-common
Continues the client.utils dissolution from the hudi-client-common
reorg, dispersing hudi-common's org.apache.hudi.util by function:

- HoodieFileMetadataMerger -> org.apache.hudi.core.io.storage: it
  merges file footer metadata (bloom filters, column ranges) for the
  parquet binary-copy write path, whose contracts live there; its test
  moves alongside, finally matching its package.
- Lazy -> org.apache.hudi.common.util (33 consumers, pure substrate).
- LazyConcatenatingIterator -> org.apache.hudi.common.util.collection,
  and the three engine-agnostic iterators recently consolidated into
  client-common's org.apache.hudi.util (LazyIterableIterator,
  ConcatenatingIterator, CloseableConcatenatingIterator) are pulled up
  into hudi-common's org.apache.hudi.common.util.collection with them,
  joining the ClosableIterator family; TestConcatenatingIterator moves
  to hudi-common accordingly. OperationConverter (JCommander-bound) and
  CommonClientUtils (client-domain) remain in client-common.
- PartitionPathFilterUtil deleted: a single static predicate with two
  callers, now a public helper on HoodieTableMetadataUtil (it encodes
  metadata-table partition-path semantics); callers in
  HoodieBackedTableMetadata and CatalogBackedTableMetadata updated and
  its tests ported.
- Transient deleted: zero references repo-wide.

BREAKING CHANGE: source/binary break for externally compiled code
importing these utilities (all shipped through 1.2.0 at the old FQNs);
fails loudly, fixed by import changes. None are config values,
reflection targets, or Kryo-registered. hudi-trino-plugin untouched
(builds against released Hudi).

Verified across common, hadoop-common, client-common, spark client and
datasource, utilities (spark4.1) and flink client/datasource
(flink2.1); merger, iterator, timeline client and metadata predicate
suites pass.
