+++ hudi-common/src/main/java/org/apache/hudi/common/util/Lazy.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.util;
+package org.apache.hudi.common.util;
 
 import lombok.Getter;
