+++ hudi-common/src/main/java/org/apache/hudi/common/util/collection/CloseableConcatenatingIterator.java
@@ -7,7 +7,7 @@
  * "License"); you may not use this file except in compliance
  * with the License.  You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ * http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -16,49 +16,25 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.util;
-
-import org.apache.hudi.common.util.ValidationUtils;
+package org.apache.hudi.common.util.collection;
 
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
-import java.util.Queue;
 
 /**
- * Provides iterator interface over List of iterators. Consumes all records from first iterator element
+ * Provides closeable iterator interface over list of iterators. Consumes all records from first iterator element
  * before moving to next iterator in the list. That is concatenating elements across multiple iterators.
- *
- * @param <T>
  */
-public class ConcatenatingIterator<T> implements Iterator<T> {
+public class CloseableConcatenatingIterator<T> extends ConcatenatingIterator<T> {
 
-  private final Queue<Iterator<T>> allIterators;
-
-  public ConcatenatingIterator(List<? extends Iterator<T>> iterators) {
-    allIterators = new LinkedList<>(iterators);
+  public CloseableConcatenatingIterator(List<ClosableIterator<T>> iterators) {
+    super(iterators);
   }
 
   @Override
-  public boolean hasNext() {
-    while (!allIterators.isEmpty()) {
-      if (allIterators.peek().hasNext()) {
-        return true;
-      }
-      // iterator at current head is done. move ahead
-      advanceIterator();
-    }
-
-    return false;
-  }
-
   protected Iterator<T> advanceIterator() {
-    return allIterators.poll();
-  }
-
-  @Override
-  public T next() {
-    ValidationUtils.checkArgument(hasNext(), "No more elements left");
-    return allIterators.peek().next();
+    ClosableIterator<T> previous = (ClosableIterator<T>) super.advanceIterator();
+    previous.close();
+    return previous;
   }
 }
