+++ hudi-common/src/main/java/org/apache/hudi/common/util/collection/LazyIterableIterator.java
@@ -16,7 +16,7 @@
  * limitations under the License.
  */
 
-package org.apache.hudi.util;
+package org.apache.hudi.common.util.collection;
 
 import java.util.Iterator;
