+++ hudi-hadoop-common/src/test/java/org/apache/hudi/parquet/io/TestHoodieParquetFileBinaryCopierPrefetch.java
@@ -18,8 +18,8 @@
 
 package org.apache.hudi.parquet.io;
 
+import org.apache.hudi.core.io.storage.HoodieFileMetadataMerger;
 import org.apache.hudi.storage.StoragePath;
-import org.apache.hudi.util.HoodieFileMetadataMerger;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
